# S3 Select: optimize output