--- a/pkg/s3select/select.go
+++ b/pkg/s3select/select.go
@@ -340,7 +340,10 @@ func (s3Select *S3Select) marshal(buf *bytes.Buffer, record sql.Record) error {
 		if err != nil {
 			return err
 		}
-
+		err = bufioWriter.Flush()
+		if err != nil {
+			return err
+		}
 		buf.Truncate(buf.Len() - 1)
 		buf.WriteString(s3Select.Output.CSVArgs.RecordDelimiter)