--- a/pkg/s3select/sql/parser.go
+++ b/pkg/s3select/sql/parser.go
@@ -103,6 +103,9 @@ type JSONPathElement struct {
 type JSONPath struct {
 	BaseKey  *Identifier        `parser:" @@"`
 	PathExpr []*JSONPathElement `parser:"(@@)*"`
+
+	// Cached values:
+	pathString string
 }
 
 // AliasedExpression is an expression that can be optionally named