--- a/pkg/s3select/sql/record.go
+++ b/pkg/s3select/sql/record.go
@@ -43,6 +43,11 @@ type Record interface {
 	WriteCSV(writer io.Writer, fieldDelimiter rune) error
 	WriteJSON(writer io.Writer) error
 
+	// CopyFrom will copy all records from the incoming and append them to the existing records.
+	// The source record must be of a similar type as destination.
+	CopyFrom(src Record) error
+	Reset()
+
 	// Returns underlying representation
 	Raw() (SelectObjectFormat, interface{})