--- a/pkg/s3select/sql/statement.go
+++ b/pkg/s3select/sql/statement.go
@@ -219,11 +219,11 @@ func (e *SelectStatement) AggregateRow(input Record) error {
 
 // Eval - evaluates the Select statement for the given record. It
 // applies only to non-aggregation queries.
-func (e *SelectStatement) Eval(input, output Record) (Record, error) {
+func (e *SelectStatement) Eval(input, output Record) error {
 	ok, err := e.isPassingWhereClause(input)
 	if err != nil || !ok {
 		// Either error or row did not pass where clause
-		return nil, err
+		return err
 	}
 
 	if e.selectAST.Expression.All {
@@ -234,14 +234,13 @@ func (e *SelectStatement) Eval(input, output Record) (Record, error) {
 		if e.limitValue > -1 {
 			e.outputCount++
 		}
-
-		return input, nil
+		return output.CopyFrom(input)
 	}
 
 	for i, expr := range e.selectAST.Expression.Expressions {
 		v, err := expr.evalNode(input)
 		if err != nil {
-			return nil, err
+			return err
 		}
 
 		// Pick output column names
@@ -259,7 +258,7 @@ func (e *SelectStatement) Eval(input, output Record) (Record, error) {
 		e.outputCount++
 	}
 
-	return output, nil
+	return nil
 }
 
 // LimitReached - returns true if the number of records output has