# Description

* Queue output items and reuse them.
* Remove the unneeded type system in sql and just use the Go type system.
* Add a 1 item capacity to the writer output.
* Prealloc a reasonable output buffer.

In best case this is more than an order of magnitude speedup:

```
>> SELECT * FROM input:

Before #8200:
BenchmarkSelectAll_1M-12    	       1	1841049400 ns/op	274299728 B/op	 4198522 allocs/op
After #8200:
BenchmarkSelectAll_1M-12    	       1	1417212000 ns/op	   0.74 MB/s	153457904 B/op	 2172063 allocs/op
With this:
BenchmarkSelectAll_1M-12    	      10	 106412080 ns/op	   9.85 MB/s	93199761 B/op	 1119696 allocs/op

13x speedup

>> SELECT id FROM input:
Before:
BenchmarkSingleCol_1M-12    	       1	2051519400 ns/op	   0.49 MB/s	336607936 B/op	14094968 allocs/op
After:
BenchmarkSingleCol_1M-12    	       9	 119898233 ns/op	   8.34 MB/s	126870314 B/op	 5074254 allocs/op

17x speedup
```


The MB/s is rows/second.

## Motivation and Context

Faster S3 select.

## How to test this PR?

There are tests in place and this should not alter any functionality.

## Types of changes
- [x] Optimization

## Checklist: